Remove broken git alias
This was a useful git alias that I remember using once, but for some reason it has broken. Probably happened after an update to Git, but I can’t nail it down. I only used it the one time, so I am going to remove it for now. I can always bring it back later.

Closes #20